finale version

+++ src/app/application/list-utilisateur/list-utilisateur.component.ts
@@ -3,6 +3,8 @@ import { Subscription } from 'rxjs';
 import { NotificationService } from 'src/app/services/notification.service';
 import { UtilisateurService } from 'src/app/services/utilisateur.service';
 import { IUser } from 'src/app/_shared/Models/i-user';
+import Swal from 'sweetalert2';
+
 
 @Component({
   selector: 'app-list-utilisateur',
@@ -12,11 +14,12 @@ import { IUser } from 'src/app/_shared/Models/i-user';
 export class ListUtilisateurComponent implements OnInit {
   private subs = new Subscription();
   listUtilisateur: IUser[] =[];
-
   constructor(private utilisateurService : UtilisateurService,
-    private notificationService : NotificationService) { }
+    private notificationService : NotificationService,
+   ) { }
 
   ngOnInit(): void {
+   
     this.getUsers()
   }
 
@@ -37,13 +40,24 @@ export class ListUtilisateurComponent implements OnInit {
 
 
   deleteUser(idUSer : any){
-    this.subs.add(
-      this.utilisateurService.deleteUserbyId(idUSer).subscribe((response)=>{
-        console.log(response)
-        this.notificationService.error("Supprission",  "Utilisateur Supprimé Avec Succès")
-        this.getUsers();
-      })
-    )
+    Swal.fire({
+      title: 'Veuillez confirmer cette action!',
+      text: "Vous ne pourrez pas récupérer l'utilisateur",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Oui!',
+      cancelButtonText: 'Non',
+    }).then((result) => {
+      if (result.value) {
+        this.subs.add(
+          this.utilisateurService.deleteUserbyId(idUSer).subscribe((response)=>{
+            console.log(response)
+            this.notificationService.error("Supprission",  "Utilisateur Supprimé Avec Succès")
+            this.getUsers();
+          })
+        )
+      }
+    });
   }
 
   ngOnDestroy(): void {
